In my new formats (v4norm and v4root), do *not* store the file contents of files which are hard links twice - instead, hard links' file sizes are set to 0. This goes along with forward link table creation, and sort of clones the chk_lnk behaviour for tar/ustar but differently (including inode number anonymisation).
This does not matter because on extraction, no file is written, instead
a hard link is created. If hard links are on different filesystems during
extraction, an error is generated and no file is being written - same as
with regular cpio/sv4cpio/sv4crc.

This change does not affect the other file formats, because I do not
have access to the SV4CPIO and POSIX CPIO specifications, respectively
cannot easily see whether it breaks the specs.

WFM.